[IDLE-114] refactor: API 공통 관리 및 패키지 구조 리팩터링 - /data/network의 api 파일에서 baseUrl 및 API 명세 관리 - 패키지 구조 리팩터링   - data layer: model, network, repository 관리   - ui layer: view, viewmodel 관리   - res: color, string 등 공통 상수 관리   - utils: logic, animation 등 유틸 관리   - ui layer는 큰 기능 단위 > 세부 기능 단위로 분리
