+++ frontend/lib/res/colors/auth_color.dart
