+++ frontend/lib/ui/auth/components/logo.dart
