+++ frontend/lib/ui/auth/components/sign_in_input.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:mybrary/constants/color.dart';
+import 'package:mybrary/res/colors/auth_color.dart';
 
-class LoginInput extends StatelessWidget {
+class SignInInput extends StatelessWidget {
   final String hintText;
   final String initialValue;
   final bool? isEnabled;
   final FormFieldSetter<String> onSaved;
   final FormFieldValidator<String> validator;
 
-  const LoginInput({
+  const SignInInput({
     required this.hintText,
     required this.validator,
     required this.initialValue,
